move UserInfo unit test to UserAvatar
